+++ src/Structure/Content/ContentTypeInterface.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Defines an interface for ContentType objects.
+ *
+ * @author  Jeremy Ward <jeremy.ward@webdevstudios.com>
+ * @package WebDevStudios\OopsWP\Structure\Content
+ * @since 2019-04-01
+ */
+
+namespace WebDevStudios\OopsWP\Structure\Content;
+
+use WebDevStudios\OopsWP\Utility\Registerable;
+
+/**
+ * Interface ContentTypeInterface
+ *
+ * @author  Jeremy Ward <jeremy.ward@webdevstudios.com>
+ * @package WebDevStudios\OopsWP\Structure\Content
+ * @since   2019-04-01
+ */
+interface ContentTypeInterface extends Registerable {
+
+}