+++ src/Structure/Plugin/PluginInterface.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Defines a contract for WordPress plugin's main class file registration.
+ *
+ * @author  Jeremy Ward <jeremy.ward@webdevstudios.com>
+ * @package WebDevStudios\OopsWP\Structure\Plugin
+ * @since 2019-04-01
+ */
+
+namespace WebDevStudios\OopsWP\Structure\Plugin;
+
+use WebDevStudios\OopsWP\Utility\Runnable;
+
+/**
+ * Interface PluginInterface
+ *
+ * @author  Jeremy Ward <jeremy.ward@webdevstudios.com>
+ * @package WebDevStudios\OopsWP\Structure\Plugin
+ * @since   2019-04-01
+ */
+interface PluginInterface extends Runnable {
+
+}