+++ src/Utility/Hookable.php
@@ -19,7 +19,7 @@ interface Hookable {
 	/**
 	 * Register actions and filters with WordPress.
 	 *
-	 * @since 1.0.0
+	 * @since 0.1.0
 	 */
 	public function register_hooks();
 }