+++ src/Utility/Runnable.php
@@ -21,7 +21,7 @@ interface Runnable {
 	/**
 	 * Run the initialization process.
 	 *
-	 * @since 1.0.0
+	 * @since 0.1.0
 	 */
 	public function run();
 }